Integrate generate_thrift.sh as source generator
Integrate the generate_thrift.sh as a custom task and add it to
`sourceGenerators` sbt setting. This way, sbt manages task execution
and makes sure that task's generated source files are properly picked
up by javac.

I had to put a bit of magic to make the task cacheable so we rerun the
source generation only when thrift definitions or the bash script itself
has changed.