Pointer, Offset and ByteCount abstractions in mmalloc
Introudce Pointer, Offset and ByteCount abstractions to mmalloc
implementation and its APIs. The motivation is the not use naked Ints
and Longs for distinct concepts like capacity, memory offset or memory
pointer. By introducing wrapper classes, we can track these things more
explicitly and separately.

The direct motivation for this change was an integer overflow bug that
was caused by mixing Ints and Longs in an expression that was computing
a pointer based of an offset.

This commit fixes two integer overflow bugs that I found.